prod/requirements.txt
torch
torchvision
ultralytics
streamlit
opencv-python
Pillow
matplotlib
pandas
